# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Enterprise Real Estate ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /installment-plans:
    get:
      operationId: listInstallmentPlans
      tags: [installmentPlans]
      summary: List installment-plans
      parameters:
        - { name: page, in: query, required: false, schema: { type: integer } }
        - { name: pageSize, in: query, required: false, schema: { type: integer } }
        - { name: search, in: query, required: false, schema: { type: string } }
        - { name: companyId, in: query, required: false, schema: { type: string } }
        - { name: contractId, in: query, required: false, schema: { type: string } }
        - { name: frequency, in: query, required: false, schema: { type: string } }
        - { name: status, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: InstallmentPlans
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstallmentPlanListResponse"
    post:
      operationId: createInstallmentPlan
      tags: [installmentPlans]
      summary: Create a InstallmentPlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstallmentPlanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstallmentPlan"
  /installment-plans/{id}:
    get:
      operationId: getInstallmentPlan
      tags: [installmentPlans]
      summary: Get a InstallmentPlan
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: InstallmentPlan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstallmentPlan"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateInstallmentPlan
      tags: [installmentPlans]
      summary: Update a InstallmentPlan
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstallmentPlanUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstallmentPlan"
    delete:
      operationId: deleteInstallmentPlan
      tags: [installmentPlans]
      summary: Soft-delete a InstallmentPlan
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /installment-schedules:
    get:
      operationId: listInstallmentSchedules
      tags: [installmentSchedules]
      summary: List installment-schedules
      parameters:
        - { name: page, in: query, required: false, schema: { type: integer } }
        - { name: pageSize, in: query, required: false, schema: { type: integer } }
        - { name: search, in: query, required: false, schema: { type: string } }
        - { name: companyId, in: query, required: false, schema: { type: string } }
        - { name: planId, in: query, required: false, schema: { type: string } }
        - { name: status, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: InstallmentSchedules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstallmentScheduleListResponse"
    post:
      operationId: createInstallmentSchedule
      tags: [installmentSchedules]
      summary: Create a InstallmentSchedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstallmentScheduleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstallmentSchedule"
  /installment-schedules/{id}:
    get:
      operationId: getInstallmentSchedule
      tags: [installmentSchedules]
      summary: Get a InstallmentSchedule
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: InstallmentSchedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstallmentSchedule"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateInstallmentSchedule
      tags: [installmentSchedules]
      summary: Update a InstallmentSchedule
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstallmentScheduleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstallmentSchedule"
    delete:
      operationId: deleteInstallmentSchedule
      tags: [installmentSchedules]
      summary: Soft-delete a InstallmentSchedule
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /installment-collections:
    get:
      operationId: listInstallmentCollections
      tags: [installmentCollections]
      summary: List installment-collections
      parameters:
        - { name: page, in: query, required: false, schema: { type: integer } }
        - { name: pageSize, in: query, required: false, schema: { type: integer } }
        - { name: search, in: query, required: false, schema: { type: string } }
        - { name: companyId, in: query, required: false, schema: { type: string } }
        - { name: scheduleId, in: query, required: false, schema: { type: string } }
        - { name: method, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: InstallmentCollections
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstallmentCollectionListResponse"
    post:
      operationId: createInstallmentCollection
      tags: [installmentCollections]
      summary: Create a InstallmentCollection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstallmentCollectionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstallmentCollection"
  /installment-collections/{id}:
    get:
      operationId: getInstallmentCollection
      tags: [installmentCollections]
      summary: Get a InstallmentCollection
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: InstallmentCollection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstallmentCollection"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateInstallmentCollection
      tags: [installmentCollections]
      summary: Update a InstallmentCollection
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstallmentCollectionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstallmentCollection"
    delete:
      operationId: deleteInstallmentCollection
      tags: [installmentCollections]
      summary: Soft-delete a InstallmentCollection
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /penalty-rules:
    get:
      operationId: listPenaltyRules
      tags: [penaltyRules]
      summary: List penalty-rules
      parameters:
        - { name: page, in: query, required: false, schema: { type: integer } }
        - { name: pageSize, in: query, required: false, schema: { type: integer } }
        - { name: search, in: query, required: false, schema: { type: string } }
        - { name: companyId, in: query, required: false, schema: { type: string } }
        - { name: penaltyType, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: PenaltyRules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PenaltyRuleListResponse"
    post:
      operationId: createPenaltyRule
      tags: [penaltyRules]
      summary: Create a PenaltyRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PenaltyRuleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PenaltyRule"
  /penalty-rules/{id}:
    get:
      operationId: getPenaltyRule
      tags: [penaltyRules]
      summary: Get a PenaltyRule
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: PenaltyRule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PenaltyRule"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updatePenaltyRule
      tags: [penaltyRules]
      summary: Update a PenaltyRule
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PenaltyRuleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PenaltyRule"
    delete:
      operationId: deletePenaltyRule
      tags: [penaltyRules]
      summary: Soft-delete a PenaltyRule
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
  /overdue-installments:
    get:
      operationId: listOverdueInstallments
      tags: [installments]
      summary: Overdue installment schedules
      parameters:
        - { name: page, in: query, required: false, schema: { type: integer } }
        - { name: pageSize, in: query, required: false, schema: { type: integer } }
        - { name: companyId, in: query, required: false, schema: { type: string } }
        - { name: planId, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: Overdue schedules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstallmentScheduleListResponse"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string }
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
    OkResponse:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
...
    InstallmentPlan:
      type: object
      required: [id, companyId, code, contractId, totalAmount, numberOfInstallments, frequency, startDate, status, isActive, createdAt]
      properties:
        id: { type: string }
        companyId: { type: string }
        code: { type: string }
        contractId: { type: string }
        totalAmount: { type: string }
        downPayment: { type: ["string", "null"] }
        numberOfInstallments: { type: integer }
        frequency: { type: string }
        startDate: { type: string }
        status: { type: string }
        isActive: { type: boolean }
        createdAt: { type: string }
    InstallmentPlanInput:
      type: object
      required: [companyId, code, contractId, startDate]
      properties:
        companyId: { type: string }
        code: { type: string }
        contractId: { type: string }
        totalAmount: { type: string }
        downPayment: { type: string }
        numberOfInstallments: { type: integer }
        frequency: { type: string }
        startDate: { type: string }
        status: { type: string }
    InstallmentPlanUpdate:
      type: object
      properties:
        companyId: { type: string }
        code: { type: string }
        contractId: { type: string }
        totalAmount: { type: string }
        downPayment: { type: string }
        numberOfInstallments: { type: integer }
        frequency: { type: string }
        startDate: { type: string }
        status: { type: string }
    InstallmentPlanListResponse:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/InstallmentPlan" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    InstallmentSchedule:
      type: object
      required: [id, companyId, planId, installmentNumber, dueDate, amount, paidAmount, status, isActive, createdAt]
      properties:
        id: { type: string }
        companyId: { type: string }
        planId: { type: string }
        installmentNumber: { type: integer }
        dueDate: { type: string }
        amount: { type: string }
        paidAmount: { type: string }
        status: { type: string }
        isActive: { type: boolean }
        createdAt: { type: string }
    InstallmentScheduleInput:
      type: object
      required: [companyId, planId, dueDate]
      properties:
        companyId: { type: string }
        planId: { type: string }
        installmentNumber: { type: integer }
        dueDate: { type: string }
        amount: { type: string }
        paidAmount: { type: string }
        status: { type: string }
    InstallmentScheduleUpdate:
      type: object
      properties:
        companyId: { type: string }
        planId: { type: string }
        installmentNumber: { type: integer }
        dueDate: { type: string }
        amount: { type: string }
        paidAmount: { type: string }
        status: { type: string }
    InstallmentScheduleListResponse:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/InstallmentSchedule" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    InstallmentCollection:
      type: object
      required: [id, companyId, scheduleId, amount, collectionDate, method, isActive, createdAt]
      properties:
        id: { type: string }
        companyId: { type: string }
        scheduleId: { type: string }
        amount: { type: string }
        collectionDate: { type: string }
        method: { type: string }
        reference: { type: ["string", "null"] }
        userId: { type: ["string", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string }
    InstallmentCollectionInput:
      type: object
      required: [companyId, scheduleId, collectionDate]
      properties:
        companyId: { type: string }
        scheduleId: { type: string }
        amount: { type: string }
        collectionDate: { type: string }
        method: { type: string }
        reference: { type: string }
        userId: { type: string }
    InstallmentCollectionUpdate:
      type: object
      properties:
        companyId: { type: string }
        scheduleId: { type: string }
        amount: { type: string }
        collectionDate: { type: string }
        method: { type: string }
        reference: { type: string }
        userId: { type: string }
    InstallmentCollectionListResponse:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/InstallmentCollection" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
...
    PenaltyRuleInput:
      type: object
      required: [companyId, code, name, nameAr]
      properties:
        companyId: { type: string }
        code: { type: string }
        name: { type: string }
        nameAr: { type: string }
        daysAfterDue: { type: integer }
        penaltyType: { type: string }
        penaltyValue: { type: string }
    PenaltyRuleUpdate:
      type: object
      properties:
        companyId: { type: string }
        code: { type: string }
        name: { type: string }
        nameAr: { type: string }
        daysAfterDue: { type: integer }
        penaltyType: { type: string }
        penaltyValue: { type: string }
    PenaltyRuleListResponse:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/PenaltyRule" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
...